requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
ppi-client